+++ exp_by_day/10.22/exp09b:new_dist_rope.sh
@@ -1,0 +1,9 @@
+## 解决viewmats的bug
+## 做了normalization
+## 加入gradient clip
+## 5帧降采样
+## resume ckpt /data1/lcy/projects/ReCamMaster/wandb/10-07-210247_Exp07g/checkpoints/step1312.ckpt
+
+
+
+bash ./scripts/train.sh -F 5 -T  -b 8 -c "2,3,4,5,6,7" -g 42 -w Exp09b